Fix relative directory error in require
If requiring a file by relative path from a file required by relative
path caused some issues where the current directory would still be
treated as the directory of the root file. This has been changed to now
always treat the directory containing the required file as the current
directory.